feat: file upload feature updated; design and functionality
